refactor: more mcts optimizations
implements a monte-carlo optimization described
[here](https://www.gameaipro.com/GameAIPro3/GameAIPro3_Chapter28_Pitfalls_and_Solutions_When_Using_Monte_Carlo_Tree_Search_for_Strategy_and_Tactical_Games.pdf),
specifically from section '28.6 Selection'.

Instead of calculating ucb1 scores for every option, keep options in a
sorted list after the initial minimum visits. This greatly reduces the
number of ucb1 calculations needed since the list being sorted means we
only need to move the first value to it's new place, which is most of
the time also near the beginning of the list.

Results in a 6% gain in monte-carlo iterations observed for a 25s search on my
hardware. More gains will be seen the larger the tree is as
a larger tree results in more ucb1 calculations relative to expansions.
Expansions are where the majority of the time is spent.